Use last transaction date (if exists) for new transactions
Senators seem to end up doing a bunch of transactions
on the same day, or at least on nearby dates. When I was going
through and entering a bunch of report data, I found it a bit
annoying to have to type out the date each time. So instead
a newly added transaction will start with the date of the last
transaction.