Set focus when adding and removing transactions
When I've been entering a bunch of these reports, its nice
to be able to completely navigate with the keyboard. One of
the things that was making it harder to do that was focus.
When adding a new transaction, its helpful if the focus goes
right to the first field (the transaction date) in the new
transaction. Similarly, if you remove a transaction, its
helpful for the focus to be restored to the add transaction
button.